### 2017-12-01 #164 @rorynolan
`ijtiff`
Correctly import TIFF files that were saved from _ImageJ_ and write TIFF files than can be correctly read by _ImageJ_. Full support for TIFF files with floating point (real-numbered) pixels. Also supports text image I/O.

```
Package: ijtiff
Title: TIFF I/O for 'ImageJ' Users
Version: 0.1.0
Authors@R: c(person("Rory", "Nolan", email = "rorynoolan@gmail.com",
                    role = c("aut", "cre", "cph")),
             person("Sergi", "Padilla-Parra", email = "spadilla@well.ox.ac.uk",
                    role = "ths"))
Maintainer: Rory Nolan <rorynoolan@gmail.com>
Description: Correctly import TIFF files that were saved from 'ImageJ' and write
    TIFF files than can be correctly read by 'ImageJ'. Full support for TIFF files
    with floating point (real-numbered) pixels. Also supports text image I/O.
SystemRequirements: libtiff
Imports: checkmate,
    RSAGA,
    filesstrings (>= 2.0.0),
    Rcpp,
    tiff,
    fields,
    grDevices,
    abind
Suggests: knitr,
    covr
LinkingTo: Rcpp
```

https://www.github.com/rorynolan/ijtiff

[e.g., "data extraction, because the package parses a scientific data file format"]
Data extraction, because the package can import (correctly) a common type of TIFF file that is not currently well-supported in R.

R users who also use _ImageJ_, or who collaborate with people who use _ImageJ_.
There's another R package for TIFF I/O, but it does not interface well with _ImageJ_ (which I would argue is very important). It's also not actively maintained (I'm not getting responses when I file an issue).

- [ y] does not violate the Terms of Service of any service it interacts with.
- [ y] has a CRAN and OSI accepted license.
- [ y] contains a README with instructions for installing the development version.
- [ y] includes documentation with examples for all functions.
- [ y] contains a vignette with examples of its essential functions and uses.
- [ y] has a test suite.
- [ y] has continuous integration, including reporting of test coverage, using services such as Travis CI, Coeveralls and/or CodeCov.
- [ y] I agree to abide by [ROpenSci's Code of Conduct](https://github.com/ropensci/onboarding/blob/master/policies.md#code-of-conduct) during the review process and in maintaining my package should it be accepted.

- [ y] Do you intend for this package to go on CRAN?
- [ y] Do you wish to automatically submit to the [Journal of Open Source Software](http://joss.theoj.org/)? If so:
    - [y ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements).
    - [ y] The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with a high-level description in the package root or in `inst/`.
    - [ y] The package is deposited in a long-term repository with the DOI: 10.5281/zenodo.1068230
- [n ] Do you wish to submit an Applications Article about your package to [Methods in Ecology and Evolution](http://besjournals.onlinelibrary.wiley.com/hub/journal/10.1111/(ISSN)2041-210X/)? If so:
	- [ ] You intend to archive the code for the package in a long-term repository which meets the requirements of the journal.

- [ y] Does `R CMD check` (or `devtools::check()`) succeed?  Paste and describe any errors or warnings:

- [ y] Does the package conform to [rOpenSci packaging guidelines](https://github.com/ropensci/onboarding/blob/master/packaging_guide.md)? Please describe any exceptions:

Andrzej K. Oleś (aoles)
