### 2017-12-02 #164 @sckott
Thank you for your submission @rorynolan 😸

It's a good fit for rOpenSci - and we'll assign an editor asap
