### 2017-12-08 #164 @sckott
(@rorynolan is there a way to instal ImageJ on osx 10.13 (High Sierra), seems I don't need it to run your pkg, but still curious)
