### 2017-12-08 #164 @sckott
### Editor checks:

---

#### Editor comments

Thanks for your submission @rorynolan !!

Goodpractice output for you and reviewers to consider:

```r

It is good practice to

    src/common.c:21:NA
    src/common.c:22:NA
    src/common.c:59:NA
    src/common.c:60:NA
    src/common.c:61:NA
    ... and 196 more lines

  ✖ avoid calling setwd(), it changes the global environment.
    If you need it, consider using on.exit() to restore the working
    directory.

    R/write.R:37:13
    R/write.R:38:5
    tests/testthat/test_graphics.R:5:11
    tests/testthat/test_graphics.R:6:3
    tests/testthat/test_io.R:4:11
    ... and 18 more lines
```

Seeking reviewers now 🕐

---

Reviewers:  xx xx
Due date: xxx
