### 2017-12-08 #164 @rorynolan
Hi @sckott,
To install ImageJ for Mac, head to https://imagej.net/Fiji/Downloads and download Fiji. Fiji stands for "Fiji Is Just ImageJ," but what it really is is ImageJ packaged up nicely with some add-ons which is easy to install. Kind of like a version of R that comes with the `tidyverse` installed.
To any potential reviewers, with regard to the `goodpractice` output, I've covered as much of the code lines as I can, and you'll notice that wherever I've used `setwd()`, I do indeed use `on.exit()` to restore the working directory.
Thanks for your selfless work.
Rory
