### 2017-12-30 #164 @jeroen
# Review

Thank you for your submission and for introducing me to ImageJ!

ROpenSci is working on improving image processing/analysis tools in R and you are doing interesting work in this space. We should see if we can collaborate more in the future.

## Main Points

The readme explains that `ijtiff` can be used to read tiff files that were written by ImageJ and that it may also work for some other tiff files. As a user who is not familiar with ImageJ I am confused about what this means. It would be nice to elaborate about how these tiff files are different from regular tiff files:

 - Are there multiple tiff formats?
 - Are these ImageJ files still proper tiff format or some derivative format?
 - Does ijtiff read/write additional image metadata ?

Perhaps you should emphasize support for high dimensionality (multi frame) tiff images which I think that this is mainly about? Or am I missing something?

## Example

The readme example image has 2 channels of 5 frames but does not explain what each channel contains. How does the user know what color or channel type the data is? Perhaps a 'real world' visible image data would be more helpful.

If we read the example image using the 'magick' package we get 10 frames with 1 channel instead of 5 frames with 2 channels.

```r
library(magick)
path_2ch_ij <- system.file("img", "2ch_ij.tif", package = "ijtiff")
image_read(path_2ch_ij)
```

Is this a bug in ImageMagick (ignoring some tiff property) or a custom property of ImageJ to group frames into channels?

I am curious if some of ijtiff functionality could also be accomplished with the magick package? The benefit of ImageMagick is that it is not limited to tiff but will work for any image format. For example:

```r
library(magick)
path <- system.file("extdata", "bleached.tif", package = "detrendr")
img <- magick::image_read(path)
frames <- lapply(img, image_data)
```

Can you write a little bit in your readme or vignette about what additional functionality ijtiff adds?

## Small comments

I tried this example image from the ImageJ website: https://imagej.nih.gov/ij/images/Composite.tif. However it gives an error when reading this with ijtiff. Is this expected?

Why `Depends: R > 3.4` ? Unless you are using features that are only available in R 3.4 it would be nice to be less restrictive so that the package can also be used on older versions of R.

Package does not build out of the box on Windows as promised. I will send a PR to fix this :)
