### 2018-01-05 #164 @sckott
Thanks for your reviews @jeroen  and @jonclayden !

@rorynolan reviews are now in. There's no one way to respond to reviewers, but here's some examples of responses to reviewers that we like:

- https://github.com/ropensci/onboarding/issues/160#issuecomment-355043656
- https://github.com/ropensci/onboarding/issues/152#issuecomment-341228007
- https://github.com/ropensci/onboarding/issues/141#issuecomment-331675487
