### 2018-01-08 #164 @rorynolan
## Responding to @jeroen

### Main Points
* __Are there multiple tiff formats? Are these ImageJ files still proper tiff format or some derivative format?__
  - No. The README now better explains that the main difference is that ImageJ has a weird (and I would say bad) way of encoding channel information, but the TIFF files that it writes are still bona fide TIFFs, not a different/derivative format. Please read the new README section entitled _The peculiarity of ImageJ TIFF files_ and let me know if I have now succeeded in making it clear what the deal is with _ImageJ_ TIFF files.
* __Does ijtiff read/write additional image metadata ?__
  - It writes to the `TIFFTAG_SOFTWARE` field to say that the TIFF was written with `ijtiff`. I have now mentioned this in the README. It currently reads all the TIFF tags that the original `tiff` package did and sets them as attributes to the read `ijtiff_img` object. In future, I plan to include all tags listed on https://www.awaresystems.be/imaging/tiff/tifftags.html as metadata when reading, and the option to write them.
* __Perhaps you should emphasize support for high dimensionality (multi frame) tiff images which I think that this is mainly about? Or am I missing something?__
  - This is more about reading channel information correctly. The original `ijtiff` package could read multi-frame images quite effectively, so `ijtiff` is not really a step forward in that regard. Admittedly, the confusion caused by the weird way _ImageJ_ writes channel information is heightened with multi-frame images, but the point of `ijtiff` is really to deal correctly with channel information. I hope this is now clear from the new sections in the README.

### Example
* __The readme example image has 2 channels of 5 frames but does not explain what each channel contains. How does the user know what color or channel type the data is? Perhaps a 'real world' visible image data would be more helpful.__
  - I've now changed my example to the banana cheerleader in front of the R logo that you use in your `magick` package and gone to more effort to make clear what the channels and frames represent. I hope you don't mind me using it, happy to remove it if you mind.
* __If we read the example image using the 'magick' package we get 10 frames with 1 channel instead of 5 frames with 2 channels. Is this a bug in ImageMagick (ignoring some tiff property) or a custom property of ImageJ to group frames into channels?__
  - I'm sure that this is just that _ImageMagick_ doesn't account for the weird way in which _ImageJ_ (sometimes) encodes channel information in `TIFFTAG_DESCRIPTION`. Personally, I think this is more of a problem with _ImageJ_ than with _ImageMagick_.
* __I am curious if some of ijtiff functionality could also be accomplished with the magick package? The benefit of ImageMagick is that it is not limited to tiff but will work for any image format. Can you write a little bit in your readme or vignette about what additional functionality ijtiff adds?__
  - I do a lot of imaging fluorescence correlation spectroscopy (FCS) and that type of image processing. For purposes like this, it's best to have images as numeric arrays. Also, having viewed an image in _ImageJ_, I think it's the most understandable form of image representation, rather than a list whose elements are 2D or 3D arrays; for most cases, the list elements all have the same type and dimension, so the list is perfectly good to be represented as a numeric array. To get to an array with `magick`, one has to go through `image_read() %>% lapply(image_data) %>% lapply(as.numeric) %>% purrr::reduce(~ abind(., along = 4)` which is OK, but a bit of a pain compared to `read_tif()`. Furthermore, like `tiff` and `EBImage`, `magick` reads the pixel values scaled to [0, 1], which is not the native TIFF representation. Some analyses like FCS are not scale-invariant, so this scaling needs to be undone to proceed. I prefer to hand the user the data in its native representation and to let them scale it if they want.
In short, the functionality that `ijtiff` adds over `magick` is that is doesn't miss the weirdly encoded channel information from _ImageJ_ and it gives the _ImageJ_ user who is beginning to use R their data in the format they're most likely to expect and want. I hope that this is now clear from the README and vignette.

## Small Comments

* __I tried this example image from the ImageJ website: https://imagej.nih.gov/ij/images/Composite.tif. However it gives an error when reading this with `ijtiff`. Is this expected?__
  - You've uncovered a weird one here. The error is `TIFFFillStrip: Invalid strip byte count 0, strip 1`. This is saying that `TIFFTAG_STRIPBYTECOUNTS` is either missing or zero (I can't tell which), but using _ImageJ_, I can open the image and I can see that `TIFFTAG_STRIPBYTECOUNTS` is equal to 143640. This makes sense because the image is 171x140 pixels, 3 channels and 16-bit (or 2-byte) and 171*140*3*2==143640. However all of `ijtiff`, `tiff` and `magick` error when trying to read this image which tells me that somewhere within `TIFFFillStrip`, `libtiff` is failing to figure our that `TIFFTAG_STRIPBYTECOUNTS` has value 143640. It's beyond me to fix this. With `ijtiff`, I can correctly read Images that `tiff` and `magick` read incorrectly, but I'm not able to read images that they can't read at all.

* __Why Depends: R > 3.4 ? Unless you are using features that are only available in R 3.4 it would be nice to be less restrictive so that the package can also be used on older versions of R.__
  - This was just me being silly. This is fixed now.

* __Make sure you add the missing PROTECT() statements from: https://raw.githubusercontent.com/kalibera/cran-checks/master/rchk/results/ijtiff.out__
  - Done.

* __Package does not build out of the box on Windows as promised. I will send a PR to fix this :)__
  - This is another weird one. Before (when I just put the onus on the user to configure `libtiff` before installing `ijtiff`), the package worked fine on CRAN. Now (after your kind PR), it works on 64-bit windows on CRAN, but not on 32-bit. It just crashes when `read_tif()` is called in examples and tests. AppVeyor suffers the same issue as 32-bit CRAN. I need help to get around this, I can't think of how to fix it. The build errors are not informative, in fact they're blank.

### Finally
* __ROpenSci is working on improving image processing/analysis tools in R and you are doing interesting work in this space. We should see if we can collaborate more in the future.__
I'm delighted that you find my work interesting and I'm very keen to help, contribute and collaborate in any capacity.

Thank you for your review and for going to the considerable bother of a PR. I really appreciate it and I think the package is already much improved thanks to you.

PS @jonclayden, currently working on a response yo your comments.
