### 2018-01-08 #164 @rorynolan
## Responding to @jonclayden

1. __The vignette is quite bare-bones. In particular, it doesn't set the scene, explaining why ImageJ generates TIFF files of the kind that it does, and why it is important to interpret them correctly in R. I would also suggest briefly explaining channels and frames in general, and (if appropriate) ImageJ's usage of them in particular. I would expect that "text images" may not be familiar to all readers—they weren't to me!—so perhaps you could also explain what they are.__
  - I've now addressed this in the _The Peculiarity of ImageJ TIFF files_, _Frames and Channels in TIFF files_ and _Text Images_ sections of the vignette.
2. __At 71% overall, test coverage is respectable, although it is 42% for common.c and 59% for read.c, which perhaps could be improved.__
  - A lot of the code in `read.c` and `common.c` is taken straight from Simon Urbanek's `tiff` package. There's a lot of code in there that I can't quite get my head around and that I can't see how to test (in my tests I read and write all kinds of images). However I feel it's safer to leave it in there than to remove it, as it probably caters for some edge cases that I have not foreseen. It's a shame that Urbanek's `tiff` package isn't unit tested.
3. __The tests of write_tif()'s error handling take a long time to run, presumably because of the need to generate a very large vector. Is there any way around this?__
  - This test is unnecessary so I've just removed it.
4. __The need for using setwd() isn't clear to me—although, admittedly, on.exit() is used to ensure that the working directory is always reverted. In the tests, it seems like it would be sufficient to use tempfile()s, which will be cleaned up by R. In write_tif() I note the comment that, "I've noticed that write_tif() sometimes fails when writing to far away directories". I'm happy to try and help get to the bottom of that if I can.__
  - I've removed all `setwd()`s from the tests in favour of `tempfile()`s. In `write_tif()`, when trying to write to a far away directory e.g. `a/b.tif`, I would get an error message like `cannot write to 'a/b.tif'`. The use of `setwd()` completely fixes this (for reasons unknown to me). I would appreciate your help on this, however, given that the workaround works, it's not a priority. More important is to get the package working on 32-bit Windows on CRAN (see my comment to @jeroen above regarding his PR).
5. __The ijtiff_img() function partially duplicates the functionality of the base function structure(), which sets a series of attributes on its first argument. I would suggest use of the latter to simplify the function.__
  - Thanks, this is now done. I never knew about the `structure()` function.
6. __In display(), it would be more transparent, and probably more idiomatic, to specify default pass-through parameters explicitly in the function arguments, viz. display <- function(img, axes = FALSE, col = grDevices::grey.colors(999,0,1), ...).__
  - This is now done, although I still have `axes` hidden from the user because the type of axes that `fields::image.plot()` offers can be confusing as they reverse the indices of the pixels on the y axis. I don't want to have an easy option of `axes = TRUE` because that's more trouble than it's worth.
7. __Is there a reason that namespace notation is used for required packages that are used everywhere, like checkmate? Would it not be cleaner just to import the package namespace? In my view using explicit namespace notation only really makes sense for suggested (rather than required or imported) dependencies.__
  - I was following the advice in http://r-pkgs.had.co.nz/namespace.html#imports. I like it this way because I don't have to go to the bother of `library(checkmate)` to step through a function that uses functions from `checkmate` line-by-line.

Thanks a lot for all your comments and suggestions @jonclayden. I really appreciate the time you're putting into this.
