### 2018-01-08 #164 @jonclayden
Thanks, Rory, for the quick response. I think the vignette is much clearer and more helpful now. The namespace syntax is really a matter of taste, so if you prefer it as it is then that's perfectly fine. (Hadley slightly contradicts himself, though, as he says to "always" use `::`, and then two paragraphs later says, "If you are using functions repeatedly, you can avoid `::` by importing the function with `@importFrom pkg fun`." 🤔)

The working directory issue is odd. Thanks for clarifying what you mean by a far-away directory, but I don't seem to be able to replicate the problem locally (with the relevant part of `write_tif()` commented out). Perhaps it is specific to certain paths (maybe those containing spaces somewhere?) or platforms. Unfortunately I don't know much about building packages on Windows, so I can't add much to the discussion on that front.

It's great to see more interest in image processing in R. Thanks again for your contribution!
