### 2018-01-08 #164 @sckott
(FYI, jeroen is on vacation for another two weeks, so likely won't get a response until then)
