### 2018-02-12 #164 @jeroen
Sorry for the delay. The updated readme is a big improvement, it's more clear now what problem exactly the package is solving. I think it's good to go.

I'll see if I can figure out what's going on with my 32bit windows libs.
