### 2018-02-19 #164 @sckott
is the 32bit windows libs thing something we should wait on here, or something for later?
