### 2018-02-19 #164 @jeroen
The author has already pushed to cran, it's fine.
