### 2018-02-20 #164 @rorynolan
Yeah sorry for jumping the gun and pushing to CRAN. The reason was that the old package was throwing some CRAN check errors and the CRAN maintainers were bugging me to fix them ASAP.
Regarding the 32bit libs, I think it's fine. It builds automatically on 64-bit windows and 32-bit users have to install `libtiff` themselves before installing `ijtiff`, which is not a big hassle. If it's not an easy fix, I'm sure @jeroen's talents would be better spent elsewhere.
Thanks for all your work on this @jeroen, I'm really grateful for it, very selfless of you.
Also thanks @sckott for making all of this happen. This is the most helpful review process I've ever experienced.
