### 2018-02-26 #164 @rorynolan
Thank you @sckott for a wonderful review experience!
* I have transferred the project to `ropensci` and done all the stuff with links and badges and the `.github` stuff.
* I am interested in writing a longer form blog post 👍

Finally, can you tell me how the JOSS paper proceeds from here?
Thanks again!
Rory
