### 2018-02-27 #164 @sckott
Great, thanks @rorynolan and glad you enjoyed it!

For JOSS:

* I'ved added a Zenodo webhoook to your repo. Generate a new release associated with a git tag in your repo. This doesn't mean you have to submit to CRAN with that git tag/release, its just to get a Zenodo DOI - and you can put a zenodo badge in your readme
* Then submit your `paper.md` to JOSS at http://joss.theoj.org/papers/new
