### 2018-02-27 #164 @stefaniebutland
@rorynolan Glad to hear the review experience worked well for you and doubly glad you're interested in writing a post.

Here are some editorial and technical guidelines: https://github.com/ropensci/roweb2#contributing-a-blog-post. Here are some other posts in the onboarding series https://ropensci.org/tags/review/

We ask for a draft to be submitted by pull request a week prior to the publication date so that we have a chance to give you feedback. What do you think of March 20 for draft submission? If you submit earlier, it does give us an opportunity to post earlier if someone else postpones.
