### 2018-03-07 #164 @rorynolan
Hi @sckott,
I'll get to that JOSS stuff in the next few weeks. Is there any rush on it?
Thanks again for everything,
I've signed up as an ROpenSci reviewer to give something back!
Rory
