### 2018-03-07 #164 @sckott
> Is there any rush on it?

i don't know, is there any timing to consider between approval here and submission to JOSS @noamross ?
